Cleanup docstrings and type annotations
